back/raw/tw-raw-data-parser/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==3.1
oauthlib==3.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.4
pymongo==3.11.4
python-dotenv==0.21.1